backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Machine Learning
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.44.0
joblib>=1.3.0

# AI/LLM
openai>=1.10.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
firebase-admin>=6.4.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CORS
fastapi-cors>=0.0.6

# Utilities
httpx>=0.26.0
